feat(apps-docs): adds support for open and close animations to navigations items
I realized that Astro View Transitions provides some nice lifecycle
events for working with view transitions. I am making use of the
before-preparation event to set a boolean that changes the
animation-name to "none". This effectively disables the animation on
each of the collapsible lists. Only once the user interacts with a
collapsible, will the animations be enabled again. Before this change,
there was a weird bounce effect that occured since view transitions do
not prevent animations from re-running on page load.