feat(apps-docs): creates search command menu
This is the initial implemention of the search command menu that allows
a user to search for a keyword or phrase and navigate to that page in
the docs. This is no where close to polished but it is a good start.
Some noteable things remaining are the following:

- Update the styling of all of the command menu items
- Add support for <mark> on the matching characters excerpts
- Figure out a way to make it more reuseable
- Make sure that the state is cleared once it is closed. Will need to
abstract the content of the dialog from where it is rendered
- Create a button in the sidebar that when clicked, opens the command
menu